1. update mcp version 2. fix missing tool in stdio mode
