+++ server/mcp_server_sec_agent/pyproject.toml
@@ -5,7 +5,7 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.10,<4.0"
 dependencies = [
-    "mcp[cli]>=1.12.0",
+    "mcp>=1.9.4",
     "pip>=25.0.1",
     "requests>=2.32.3",
     "volcengine-python-sdk>=3.0.1",
